base: introduce Allocator::try_alloc
This patch changes the 'Allocator' interface to the use of 'Attempt'
return values instead of using exceptions for propagating errors.

To largely uphold compatibility with components using the original
exception-based interface - in particluar use cases where an 'Allocator'
is passed to the 'new' operator - the traditional 'alloc' is still
supported. But it existes merely as a wrapper around the new
'try_alloc'.

Issue #4324

+++ repos/dde_bsd/src/drivers/audio/main.cc
@@ -357,7 +357,7 @@ class Audio_in::In
 			float const scale = 32768.0f * 2;
 
 			float * const content = p->content();
-			for (int i = 0; i < 2*Audio_in::PERIOD; i += 2) {
+			for (unsigned long i = 0; i < 2*Audio_in::PERIOD; i += 2) {
 				float sample = data[i] + data[i+1];
 				content[i/2] = sample / scale;
 			}
